crabnlp/requirements.txt
openai
pytube
transformers
PyICU
pycld2
polyglot
tiktoken
python-json-logger
iso-639
faster_whisper==0.5.0
pyannote.audio==2.1.1